move block lock logic to its own hook
