+++ src/components/PresenceUI.tsx
@@ -1,8 +1,8 @@
-import { createPortal, useEffect } from '@wordpress/element';
-import { useDispatch } from '@wordpress/data';
+import { createPortal } from '@wordpress/element';
 import { useMultiCursor } from '@/useMultiCursor';
 import AvatarList from './AvatarList';
 import { CursorState, AwarenessState } from '@/hooks/types';
+import { useBlockLocking } from '@/hooks/useBlockLocking';
 
 interface PresenceUIProps {
 	awarenessState: AwarenessState;
@@ -15,43 +15,10 @@ export const PresenceUI = ( {
 	syncAwareness,
 	currentUserId,
 }: PresenceUIProps ) => {
-	const { setLockedBlocks } = useDispatch( 'gce' );
-
 	const otherUsers = awarenessState;
 
 	useMultiCursor( currentUserId, otherUsers, syncAwareness );
-
-	// Assume user's cursors carries edit intent
-	// so lock all blocks where cursors of other users are at
-	useEffect( () => {
-		const lockedBlocks: string[] = [];
-		Object.values( otherUsers ).forEach( ( user ) => {
-			if ( ! user.cursor_state ) {
-				// eslint-disable-next-line no-console
-				console.log( '🚧 when is this happening?', user );
-				return;
-			}
-			const blockIndex =
-				'blockIndex' in user.cursor_state
-					? user.cursor_state.blockIndex
-					: user.cursor_state.blockIndexStart;
-
-			const existingBlocks = wp.data
-				.select( 'core/block-editor' )
-				.getBlocks();
-
-			const block = existingBlocks[ blockIndex ];
-			if ( block ) {
-				lockedBlocks.push( block.clientId );
-			}
-		} );
-		setLockedBlocks( lockedBlocks );
-
-		// Cleanup function to remove locks when users leave
-		return () => {
-			setLockedBlocks( [] );
-		};
-	}, [ otherUsers, setLockedBlocks ] );
+	useBlockLocking( otherUsers );
 
 	if ( ! otherUsers || Object.keys( otherUsers ).length === 0 ) {
 		return null;
